requirements-dev.txt
stestr>=3.0.0
astroid==2.5
pylint==2.7.1
black==21.11b1
qiskit-sphinx-theme>=1.6
sphinx-autodoc-typehints
jupyter-sphinx
pygments>=2.4
reno>=3.4.0
nbsphinx
qutip
ddt~=1.4.2
matplotlib>=3.3.0
